requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
PyPDF2==1.26.0
python-barcode==0.13.1
xlrd==2.0.1
